Split up user and project storage (fixes #4893)
